+++ jgclark.Summaries/src/progress.js
@@ -2,15 +2,15 @@
 //-----------------------------------------------------------------------------
 // Progress update on some key goals to include in notes
 // Jonathan Clark, @jgclark
-// Last updated for v0.4.0, 12.1.2022
+// Last updated for v0.5.0, 16.1.2022
 //-----------------------------------------------------------------------------
 
 import {
+  calcHashtagStatsPeriod,
   calcMentionStatsPeriod,
   getConfigSettings,
   getPeriodStartEndDates,
 } from './summaryHelpers'
-import { logAllEnvironmentSettings } from '../../helpers/NPdev'
 import {
   unhyphenatedDate,
 } from '../../helpers/dateTime'
@@ -18,8 +18,6 @@ import { rangeToString } from '../../helpers/general'
 
 
 //-------------------------------------------------------------------------------
-
-const progressHeading = `Progress Update`
 
 function getSelectedParaIndex(): number {
   const { paragraphs, selection } = Editor
@@ -70,85 +68,91 @@ export async function insertProgressUpdate(intervalType?: string): Promise<void>
   // Get day of week (Sunday is first day of the week) or day of month
   const dateWithinInterval = (interval === 'wtd') ? (new Date()).getDay() + 1 : (new Date()).getDate()
 
-  // // Calc hashtags stats (returns two maps)
-  // const hOutputArray = []
-  // let results = await calcHashtagStatsPeriod(fromDateStr, toDateStr)
-  // const hCounts = results?.[0]
-  // const hSumTotals = results?.[1]
-  // if (hSumTotals == null || hCounts == null) {
-  //   console.log('no hSumTotals value')
-  //   return
-  // }
 
-  // // Custom sort method to sort arrays of two values each
-  // // const sortedHCounts = new Map(
-  // //   [...(hCounts?.entries() ?? [])].sort(([key1, _v1], [key2, _v2]) =>
-  // //     key1.localeCompare(key2),
-  // //   ),
-  // // )
+  // Calc hashtags stats (returns two maps)
+  const hOutputArray = []
+  // $FlowIgnore[invalid-tuple-arity]
+  let hResults = await calcHashtagStatsPeriod(fromDateStr, toDateStr, config.progressHashtags, [])
+  const hCounts = hResults?.[0]
+  const hSumTotals = hResults?.[1]
+  if (hSumTotals == null || hCounts == null) {
+    console.log('  warning: no results from calcHashtagStatsPeriod')
+  } else {
 
-  // // First process more complex 'SumTotals', calculating appropriately
-  // for (const [key, value] of hSumTotals) {
-  //   // .entries() implied
-  //   const hashtagString = config.showAsHashtagOrMention ? key : key.slice(1)
-  //   const count = hCounts.get(key)
-  //   if (count != null) {
-  //     const total: string = value.toFixed(0)
-  //     const average: string = (value / count).toFixed(1)
-  //     hOutputArray.push(
-  //       `${hashtagString}\t${count}\t(total ${total}\taverage ${average})`,
-  //     )
-  //     hCounts.delete(key) // remove the entry from the next map, as not longer needed
-  //   }
-  // }
-  // // Then process simpler 'Counts'
-  // for (const [key, value] of hCounts) {
-  //   // .entries() implied
-  //   const hashtagString = config.showAsHashtagOrMention ? key : key.slice(1)
-  //   hOutputArray.push(`${hashtagString}\t${value}`)
-  // }
-  // // If there's nothing to report, let's make that clear, otherwise sort output
-  // if (hOutputArray.length > 0) {
-  //   hOutputArray.sort()
-  // } else {
-  //   hOutputArray.push('(none)')
-  // }
+    // Custom sort method to sort arrays of two values each
+    // const sortedHCounts = new Map(
+    //   [...(hCounts?.entries() ?? [])].sort(([key1, _v1], [key2, _v2]) =>
+    //     key1.localeCompare(key2),
+    //   ),
+    // )
 
-  // Calc mentions stats (returns two maps)
-  const mOutputArray = []
-  let results = await calcMentionStatsPeriod(fromDateStr, toDateStr)
-  const mCounts = results?.[0]
-  const mSumTotals = results?.[1]
-  if (mCounts == null || mSumTotals == null) {
-    return
-  }
+    // // First process more complex 'SumTotals', calculating appropriately
+    // for (const [key, value] of hSumTotals) {
+    //   // .entries() implied
+    //   const hashtagString = key.slice(1) // show without leading '#' to avoid double counting issues
+    //   const count = hCounts.get(key)
+    //   if (count != null) {
+    //     const total: string = value.toFixed(0)
+    //     const average: string = (value / count).toFixed(1)
+    //     hOutputArray.push(
+    //       `${hashtagString}\t${count}\t(total ${total}\taverage ${average})`,
+    //     )
+    //     hCounts.delete(key) // remove the entry from the next map, as not longer needed
+    //   }
+    // }
 
-  // First process more complex 'SumTotals', calculating appropriately
-  for (const [key, value] of mSumTotals) {
-    // .entries() implied
-    const mentionString = key.slice(1) // show without leading '@' to avoid double counting issues
-    const count = mCounts.get(key)
-    if (count != null) {
-      const total = value.toFixed(0)
-      const average = (value / count).toFixed(1)
-      mOutputArray.push(
-        `${mentionString}\t${count}\t(total ${total}\taverage ${average})`,
-      )
-      mCounts.delete(key) // remove the entry from the next map, as not longer needed
+    // Then process simpler 'Counts'
+    for (const [key, value] of hCounts) {
+      // .entries() implied
+      const hashtagString = key.slice(1) // show without leading '#' to avoid double counting issues
+      hOutputArray.push(`${hashtagString}\t${value}`)
+    }
+    // If there's nothing to report, let's make that clear, otherwise sort output
+    if (hOutputArray.length > 0) {
+      hOutputArray.sort()
+    } else {
+      hOutputArray.push('(none)')
     }
   }
-  // Then process simpler 'Counts'
-  for (const [key, value] of mCounts) {
-    const mentionString = key.slice(1) // show without leading '@' to avoid double counting issues
-    mOutputArray.push(`${mentionString}\t${value}`)
-  }
-  // If there's nothing to report, let's make that clear, otherwise sort output
-  if (mOutputArray.length > 0) {
-    mOutputArray.sort()
+  
+  // Calc mentions stats (returns two maps)
+  const mOutputArray = []
+  // $FlowIgnore[invalid-tuple-arity]
+  let mResults = await calcMentionStatsPeriod(fromDateStr, toDateStr, config.progressMentions, [])
+  const mCounts = mResults?.[0]
+  const mSumTotals = mResults?.[1]
+  if (mCounts == null || mSumTotals == null) {
+    console.log('  warning: no results from calcMentionsStatsPeriod')
   } else {
-    mOutputArray.push('(none)')
+
+    // First process more complex 'SumTotals', calculating appropriately
+    for (const [key, value] of mSumTotals) {
+      // .entries() implied
+      const mentionString = key.slice(1) // show without leading '@' to avoid double counting issues
+      const count = mCounts.get(key)
+      if (count != null) {
+        const total = value.toFixed(0)
+        const average = (value / count).toFixed(1)
+        mOutputArray.push(
+          `${mentionString}\t${count}\t(total ${total}\taverage ${average})`,
+        )
+        mCounts.delete(key) // remove the entry from the next map, as not longer needed
+      }
+    }
+    // Then process simpler 'Counts'
+    for (const [key, value] of mCounts) {
+      const mentionString = key.slice(1) // show without leading '@' to avoid double counting issues
+      mOutputArray.push(`${mentionString}\t${value}`)
+    }
+    // If there's nothing to report, let's make that clear, otherwise sort output
+    if (mOutputArray.length > 0) {
+      mOutputArray.sort()
+    } else {
+      mOutputArray.push('(none)')
+    }
   }
 
+  // Now insert the summary to the current note
   if (Editor == null) {
     console.log(`error: no note is open`)
   } else {
@@ -159,11 +163,12 @@ export async function insertProgressUpdate(intervalType?: string): Promise<void>
     }
     // console.log(`\tinserting results to current note (${Editor.filename ?? ''}) at line ${currentLineIndex}`)
     Editor.insertHeading(
-      `${progressHeading}: Day ${dateWithinInterval} for ${periodString}`, // TODO: make this from ~ String(config.mentionCountsHeading),
+      `${config.progressHeading}: Day ${dateWithinInterval} for ${periodString}`,
       currentLineIndex,
       3,
     )
     Editor.insertParagraph(mOutputArray.join('\n'), currentLineIndex + 1, 'text')
+    Editor.insertParagraph(hOutputArray.join('\n'), currentLineIndex + 1 + mOutputArray.length, 'text')
     console.log(`\tappended results to current note for day ${dateWithinInterval}.`)
   }
 }