+++ jgclark.Summaries/CHANGELOG.md
@@ -1,5 +1,10 @@
 # What's Changed in 🗃 Summaries plugin?
 
+## [0.5.0] - 2022-01-16
+### Added
+- added hashtags to the `/insertProgressUpdate` command.
+- the list of hashtags and mentions to include in Progress Updates are now specified separately, using the `progressHashtags` and `progressMentions` settings.
+
 ## [0.4.0] - 2022-01-14
 ### Added
 - added `/insertProgressUpdate` command. This writes out a summary of mentions of interest so far this week or month, showing the count/sum/average so far in that time period, to the current note. This is particularly designed to be used from a daily template by `{{insertProgressUpdate()}}`. See [README](https://github.com/NotePlan/plugins/tree/main/jgclark.Summaries/) for more details.